requirements.txt
tensorflow>=2.1.0
pytorch>=1.0.2
numpy>=1.18.1
matplotlib>=3.1.2